pkg/seq/ast: moves the package out `parser` package
