+++ pkg/sequence/parser/file_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/ufukty/diagramer/pkg/sequence/parser/match/ast"
+	"github.com/ufukty/diagramer/pkg/sequence/ast"
 )
 
 func PrintDiagram(d *ast.Diagram) {
